# Integrate latest vec_*.h from xiph/LPCNet