… src/tansig_table.h
… src/vec_avx.h