This PR integrates the latest `vec_*.h` files from [the upstream repo](https://github.com/xiph/LPCNet). As a result, we end up with the following performance improvements:

## macOS x86_64 (2019 MacBook Pro):

*`master`*:

```
HT-TM05:src mooneer$ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m1.395s
user	0m1.530s
sys	0m0.042s
HT-TM05:src mooneer$ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m1.435s
user	0m1.477s
sys	0m0.031s
HT-TM05:src mooneer$ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m1.408s
user	0m1.543s
sys	0m0.041s
HT-TM05:src mooneer$
```

=> 1.412s average

*This PR*:

```
HT-TM05:src mooneer$ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...

real	0m1.256s
user	0m1.397s
sys	0m0.032s
HT-TM05:src mooneer$ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m1.305s
user	0m1.413s
sys	0m0.040s
HT-TM05:src mooneer$ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m1.205s
user	0m1.353s
sys	0m0.028s
HT-TM05:src mooneer$
```

=> 1.255s average (~10% speedup)

## macOS aarch64 (M1 Mac Mini)

*`master`*:

```
mooneer@ubuntu-server src % time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
( sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s)  0.93s user 0.03s system 110% cpu 0.869 total
mooneer@ubuntu-server src % time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
( sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s)  0.94s user 0.03s system 110% cpu 0.871 total
mooneer@ubuntu-server src % time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
( sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s)  0.94s user 0.03s system 110% cpu 0.872 total
```

=> 0.871s average

*This PR*:

```
mooneer@ubuntu-server src % time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
( sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s)  0.69s user 0.03s system 115% cpu 0.615 total
mooneer@ubuntu-server src % time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
( sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s)  0.67s user 0.02s system 115% cpu 0.607 total
mooneer@ubuntu-server src % time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
( sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s)  0.68s user 0.03s system 115% cpu 0.611 total
```

=> 0.611s average (~30% speedup)

## Raspberry Pi 4 (aarch64)

*`master`*:

```
mooneer@raspberrypi:~/freedv-gui/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m6.648s
user	0m6.916s
sys	0m0.093s
mooneer@raspberrypi:~/freedv-gui/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m6.663s
user	0m6.987s
sys	0m0.092s
mooneer@raspberrypi:~/freedv-gui/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m6.590s
user	0m6.857s
sys	0m0.082s
mooneer@raspberrypi:~/freedv-gui/LPCNet/build/src $ 
```

=> 6.634s average

*This PR*:

```
mooneer@raspberrypi:~/freedv-gui/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m4.420s
user	0m4.672s
sys	0m0.098s
mooneer@raspberrypi:~/freedv-gui/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m4.364s
user	0m4.625s
sys	0m0.090s
mooneer@raspberrypi:~/freedv-gui/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m4.431s
user	0m4.752s
sys	0m0.078s
mooneer@raspberrypi:~/freedv-gui/LPCNet/build/src $ 
```

=> 4.405s average (~30% speedup)