@tmiw 2022-12-09
Apparently with this PR LPCNet is basically real time on the Pi 4B now. `lpcnet_enc` seems to only take 0.3-0.4s and wia.wav is 4.2s long (per Audacity).

Of course, it's *just* real time, so it still might not perform well enough without bringing in additional changes from upstream.