@drowe67 2022-12-09
Thanks @tmiw - did you listen to a few sample files to confirm it is working OK?