@tmiw 2022-12-10
> Thanks @tmiw - did you listen to a few sample files to confirm it is working OK?

Seems to work okay with wia.wav and all.wav but I don't detect much if any difference in the resulting audio (which is probably a good thing). 

However, I did notice a compiler error on the Raspberry Pi 3 but that may be because I'm running an older OS on it. Will need to confirm with a newer SD card.