@tmiw 2022-12-10
Turns out that the compiler errors were due to the new vec_* files containing some aarch64 specific code. After removing that, the Pi 3B+ gives me the following results for master:

master:

```
pi@piaware:~/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m50.840s
user	0m53.434s
sys	0m0.135s
pi@piaware:~/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m51.135s
user	0m53.744s
sys	0m0.144s
pi@piaware:~/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m51.307s
user	0m53.776s
sys	0m0.250s
pi@piaware:~/LPCNet/build/src $ 
```

=> 51.094s average

vs. this PR:

```
pi@piaware:~/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m35.547s
user	0m38.187s
sys	0m0.204s
pi@piaware:~/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m35.562s
user	0m38.273s
sys	0m0.220s
pi@piaware:~/LPCNet/build/src $ time (sox ../../wav/wia.wav -t raw -r 16000 - | ./lpcnet_enc -s | ./lpcnet_dec -s > /dev/null)
...
real	0m37.578s
user	0m39.662s
sys	0m0.560s
pi@piaware:~/LPCNet/build/src $
```

=> 36.229s average (30% improvement)